Move logic into printer
